+++ ContentPatcher/Framework/Conditions/Weather.cs
@@ -3,7 +3,7 @@ namespace ContentPatcher.Framework.Conditions
     /// <summary>An in-game weather.</summary>
     internal enum Weather
     {
-        /// <summary>The weather is sunny (includes hardcoded-sun days like festivals and weddings).</summary>
+        /// <summary>The weather is sunny (including festival/wedding days). This is the default weather if no other value applies.</summary>
         Sun,
 
         /// <summary>Rain is falling, but without lightning.</summary>
@@ -13,6 +13,9 @@ namespace ContentPatcher.Framework.Conditions
         Storm,
 
         /// <summary>Snow is falling.</summary>
-        Snow
+        Snow,
+
+        /// <summary>The wind is blowing with visible debris (e.g. flower petals in spring and leaves in fall).</summary>
+        Wind
     }
 }